Add dedicated VanillaRegistries
